added backup feature.
Problem: vdiff-viewer -d big.db --merge is very slow because small, sequential writes are really slow to sqlite (or any file-backed database). The transaction safety is not really needed here.

Solution: vdiff-viewer -d :memory: --backup big.db --merge solves this. Using the reserved filename :memory: tells vdiff to use an in-memory sqlite database. --backup causes vdiff to write the in-memory (after everything is completed) to the given "backup" file.

This is implemented in the Application scaffold and is therefore available in all vdiff applications.